requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-cors-headers==4.3.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt[crypto]==5.3.0
drf-spectacular==0.26.5
pkg_resources==0.0.0
psycopg2==2.9.5
pytz==2023.3
sqlparse==0.4.3
